pin watcher first steps

+++ rush/src/command_parser.rs
@@ -40,7 +40,7 @@ pub struct ReadCommand {
 impl Command for ReadCommand {
     fn execute<'a>(&self, fmt_buffer: &'a mut [u8], pin_manager: &mut RushPinManager) -> &'a str {
         let Id::Gpio(pin) = self.id;
-        match pin_manager.get_pin(pin as usize).to_input().read_state() {
+        match pin_manager.get_pin(pin).to_input().read_state() {
             Ok(state) => fmt_truncate(
                 fmt_buffer,
                 format_args!("state of gpio.{}: {}\n", pin, state),
@@ -81,11 +81,11 @@ pub struct WriteCommand {
 impl Command for WriteCommand {
     fn execute<'a>(&self, fmt_buffer: &'a mut [u8], pin_manager: &mut RushPinManager) -> &'a str {
         let (Id::Gpio(pin), Value::Gpio(b)) = (&self.id, &self.value);
-        match pin_manager
-            .get_pin(*pin as usize)
-            .to_output()
-            .set_state(if *b { PinState::High } else { PinState::Low })
-        {
+        match pin_manager.get_pin(*pin).to_output().set_state(if *b {
+            PinState::High
+        } else {
+            PinState::Low
+        }) {
             Ok(_) => fmt_truncate(fmt_buffer, format_args!("writing {} to gpio.{}\n", b, pin)),
             Err(err) => fmt_truncate(
                 fmt_buffer,
